Main: Handle Application instantiation failure
Command 'create' requires that a directory matching the new
project name does not already exist. Additionall 'update' and
'build' require to be run from the top-level directory of a
crosswalk-app-tools project. These cases are now checked for, and
the errors are handled, an explanatory message is printed.

Fixes https://crosswalk-project.org/jira/browse/XWALK-4089