env-vars: Move environment variable 'declaration' to backends
Backends can now 'declare' which extra environment variables they
are paying attention to, and this is printed in the help output.
Use this new functionality to move the download cache dir handling
to the android backend, so it no longer shows up in the help out-
put on other platforms.

Also update tests.

Fixes https://crosswalk-project.org/jira/browse/XWALK-4152