-pkg: New parameter -t / --targets to specify target ABIs
Ways to specify targets:
* List of ABIs, e.g. -t "armeabi-v7a x86"
* Prefix of ABI names e.g. -t "arm x86" matches armeabi-v7a as well
  as arm64-v8a, but not x86_64 because x86 is an exact match.
* By word size, e.g. -t 32 matches 32bit arm and x86, -t "32 64"
  matches all ABIs.
* When not specified, the 32-bit ABIs are built by default.

When this parameter is used together with a local crosswalk zip or
xwalk_app_template directory, only the locally available ABIs are
available to be built, missing ones will not be downloaded.

At the moment the 64bit releases are only in the newest canary
releases, so to build them, the --crosswalk=canary parameter needs
to be used, or an explicit version number that is new enough to
include the 64bit release, such as 17.45.430.0 or newer.

BUG=XWALK-5607