Android: Derive activity name from package-id
The main java activity class was always named "MainActivity" in our
built apps, but for easier automated testing with QAs tools we want
to have the name be derived from the package-id. So for example for
package-id of "com.example.foo", we now derive the activity class
name as FooActivity, using the last component case-modified, with
"Activity" suffix.

BUG=XWALK-5014