Android: Disable unused ABIs by libxwalkcore.so, not directory
The way to pick target API was to set the respective JNI libs
directory readable/unreadable. This does not work on windows,
so we always built fat APKs.

So we are changing strategy, and the skipped ABIs' libxwalkcore.so
are renamed to libxwalkcore.so.foo, such that ant does not consider
them.

BUG=XWALK-6470