lint.py: Exit with the status code of the unit tests.
We were currently invoking the Python unit test files with
`os.system()` (which is bad in itself) without bothering to check its
return code.

In practice, this meant that `lint.py` always exited with 0 regardless
of whether the unit tests it invoked passed or not.

While here, stop redirecting output on Linux to a file called "null".